requirements.txt
numpy==1.24.3
numpy-stl==3.0.1
PyOpenGL==3.1.6
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqtgraph==0.13.3
pyserial==3.5
python-utils==3.5.2
